+++ ErsatzTV.Application/Images/Queries/GetCachedImagePath.cs
@@ -3,5 +3,5 @@ using ErsatzTV.Core.Domain;
 
 namespace ErsatzTV.Application.Images;
 
-public record GetCachedImagePath(string FileName, ArtworkKind ArtworkKind, int? MaxHeight = null) : IRequest<
+public record GetCachedImagePath(string FileName, ArtworkKind ArtworkKind, string ContentType, int? MaxHeight = null) : IRequest<
     Either<BaseError, CachedImagePathViewModel>>;