+++ ErsatzTV.Application/Watermarks/WatermarkViewModel.cs
@@ -1,11 +1,12 @@
-﻿using ErsatzTV.Core.Domain;
+﻿using ErsatzTV.Application.Artworks;
+using ErsatzTV.Core.Domain;
 using ErsatzTV.FFmpeg.State;
 
 namespace ErsatzTV.Application.Watermarks;
 
 public record WatermarkViewModel(
     int Id,
-    string Image,
+    ArtworkContentTypeModel Image,
     string Name,
     ChannelWatermarkMode Mode,
     ChannelWatermarkImageSource ImageSource,